取整

+++ im-service/im-service-aggregation/src/main/java/com/starsea/im/aggregation/service/DiagnoseService.java
@@ -34,7 +34,7 @@ public interface DiagnoseService {
 
     public  List<List<Long>> getFinalRegularScore();
 
-    public  List<Double> getFinalStdScore();
+    public  List<Long> getFinalStdScore();
